% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kernel_lasso.R
\name{lasso.control}
\alias{lasso.control}
\title{lasso.control}
\usage{
lasso.control(nfolds = 10, trace.it = 1, type.measure = "auc")
}
\arguments{
\item{nfolds}{n-fold cross-validation.}

\item{trace.it}{Whether to plot the training process}

\item{type.measure}{Choose the loss funcrion.}
}
\value{
The lasso control setting
}
\description{
The same function from glmnet, which controls the training of lasso.
}
\examples{
##10-fold Cross-validation with MSE as loss function
c<-lasso.control(nfolds=10,type.measure='mse')
}
\keyword{lasso.control}
