% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Standardization.R
\name{Z_score}
\alias{Z_score}
\title{Z_score}
\usage{
Z_score(data, dataframe = FALSE)
}
\arguments{
\item{data}{Your input data, which can be numerci or data.frame}

\item{dataframe}{Wether the data is dataframe. The default is False(numeric)}
}
\description{
Z-score method is used to calculate the standardization value of data.The formula is (x-mean(x))/var(x). It can compress the data into the (0,1).
}
\examples{
##For the numeric data
data(iris,package = 'datasets')
w<-Z_score(iris[,1])
print(w)

##For the data.frame data
w1<-Z_score(iris[,-5],dataframe=TRUE)
print(w1)
}
\keyword{Z_score}
