% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kernel_lasso.R
\name{lasso.control}
\alias{lasso.control}
\title{lasso.control}
\usage{
lasso.control(nfolds = 10, trace.it = 1, type.measure = "auc")
}
\arguments{
\item{x}{Your input features, which can have to be data.frame with at least two variables.}

\item{y}{The dependent variable}

\item{sigma}{The hyperparameter of RBF kernel function, which indicates the width.}

\item{dataframe}{Wether the data is dataframe. The default is TURE}

\item{standard}{Using 'max_min_scale' or 'Z_score' method to standardize the data. NULL means no standardization}
}
\description{
Kernel_lasso is one feature selection method, which combines the feature expansion and lasso regression together. Kernel function will increase the dimensions of the existed data and then reduce the features by lasso
}
\examples{
##10-fold Cross-validation with MSE as loss function
c<-lasso.control(nfolds=10,type.measure='mse')
}
\keyword{lasso.control}
